fix some time issues...that don't occur under iPad Retina(???)

+++ date-picker-uw-ios-hw4/BDayPerson.h
@@ -18,7 +18,7 @@
 
 - (NSDate *) nextBDay;
 
-- (int) daysUntilBDay;
+- (long) daysUntilBDay;
 
 -(instancetype) initWithName:(NSString *)name
                         Date:(NSDate *)date;
